refactor(admin): group monitoring dashboard module under admin/dashboard
The overview dashboard page, its data composable and styles were stored in three unrelated folders. Fold Dashboard.vue, useDashboard.ts and dashboard.scss into one module directory.
