+++ webui/src/admin/dashboard/dashboard.scss
